Change id and _id
Now, node.id only exists if you pass an id on create. You cannot update
with an id. node._id always exists, and cannot be assigned, but it can
be queried, and returns as such (_id) always. So basically, id is just
like any other properties, except that (1) it also gets hoisted out of
`data`, and (2) you cannot update it